move to as zebra controlled shutter
